refactor: refactor update payment information
